+++ CHANGELOG.md
@@ -10,6 +10,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 - Discussion Topic title validation
 - Decisions can be rescinded
 - Heroku deploys are quicker now with a single line command
+- Add permission to view decisions
 ### Changed
 - Soft deleting questions and answers. Answers are retained even when questions are marked
   as deleted